Add files via upload
n this project I have constructed and coded (in Python) a genetic algorithm for hyperparameter tuning of CNN for image classification. With each subsequent iteration, this algorithm constructs and outputs an improved CNN architecture for a given task. Specifically in this project a handwritten digits dataset is used; however the approach and the code provided in this notebook can be easily modified for the use on the more complex datasets and classification tasks.